Added updated MAME logo, added customisable font face selection
